requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.10.1
django-crontab==0.7.1
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
